+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improved French (Français) translations - thanks to **ma203**
 - Improved Simple Chinese (中文) translations - thanks to **kur1k0**
+- Fixed an issue where the app wasn't restarting properly after language changes
 - Improved translation credits layout
 
 ## [1.0.8] - 2025-05-21